% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBoxesOther.R
\name{checkBoxesOther}
\alias{checkBoxesOther}
\title{Shiny checbox buttons with other input os last item}
\usage{
checkBoxesOther(id, label, choices = NULL, choiceNames = NULL,
  choiceValues = NULL, selected = NULL, otherLabel = "Other",
  placeholder = NULL, ...)
}
\arguments{
\item{id}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{choices}{List of values to select from (if elements of
the list are named then that name rather than the value
is displayed to the user). If this argument is provided,
then choiceNames and choiceValues must not be provided, and
vice-versa. The values should be strings; other types (such
as logicals and numbers) will be coerced to strings.}

\item{selected}{The initially selected value (if not specified then defaults to the first value)}
}
\description{
Given an input id, label, and possible choices, will
generate radio buttons. The final radio button will
be a text input field labeled as other.
}
\author{
Danny Beachnau
}
