% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extVerticalTabsetPanel.R
\name{extVerticalTabsetPanel}
\alias{extVerticalTabsetPanel}
\title{Extends options for a vertical tabset panel from the shinyWidgets library}
\usage{
extVerticalTabsetPanel(..., selected = NULL, id = NULL,
  color = "#112446", contentWidth = 9, menuSide = "left",
  addButton = F)
}
\description{
Extends options for a vertical tabset panel from the shinyWidgets library
}
\author{
DannyBeachnau
}
