% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggleButton.R
\name{toggleButton}
\alias{toggleButton}
\title{Button that iterates over options on each click}
\usage{
toggleButton(id, label, selected = NULL, choices = NULL,
  choiceNames = NULL, choiceValues = NULL)
}
\description{
Button that iterates over options on each click
}
\author{
Danny Beachnau
}
